+++ PyrusServiceDeskAndroid/pyrusservicedesk/src/main/java/com/pyrus/pyrusservicedesk/sdk/sync/CommandParamsDto.kt
@@ -11,12 +11,12 @@ internal sealed interface CommandParamsDto {
         @SerializedName("app_id") val appId: String,
         @SerializedName("comment") val comment: String?,
         @SerializedName("attachments") val attachments: List<AttachmentDataDto>?,
-        @SerializedName("ticket_id") val ticketId: Int,
+        @SerializedName("ticket_id") val ticketId: Long,
         @SerializedName("rating") val rating: Int?
     ) : CommandParamsDto
 
     data class MarkTicketAsRead(
-        @SerializedName("ticket_id") val ticketId: Int,
+        @SerializedName("ticket_id") val ticketId: Long,
         @SerializedName("user_id") val userId: String,
         @SerializedName("app_id") val appId: String,
         @SerializedName("comment_id") val commentId: Long?, // readAll if null
